server/server/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
PyMySQL==1.1.0  # ← ACTUALIZAR esta versión
cryptography==46.0.2  # ← AÑADIR este paquete
python-jose==3.3.0
passlib==1.7.4
flask-restx==1.3.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
PyJWT==2.8.0